Add IPosisBundle
This removes the need for global functions and let's the host decide
when to install certain bundles. At the same time it enforces programs
to be bundled, even if it's a single program.
Additionally this will allow for IPosisBundle-Factories if the author of
a bundle decides to implement such, making his bundle configurable and
therefore compatible with the rest of the system.

As part of this change the `queryPosisInterface` function has been moved
to IPosisProcess as well, to remove global dependencies entirely.